Make top-level class be primary
Move implementation of SiriusFactory to the top-level package, the
existing class in api.impl just wraps that class and is marked
deprecated.
